[Enhancement] Introduce GC control by cluster snapshot info (#58909)

+++ be/src/storage/lake/vacuum.cpp
@@ -33,6 +33,7 @@
 #include "storage/lake/metacache.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/lake/tablet_metadata.h"
+#include "storage/lake/tablet_retain_info.h"
 #include "storage/lake/update_manager.h"
 #include "storage/protobuf_file.h"
 #include "testutil/sync_point.h"
@@ -319,8 +320,11 @@ void run_clear_task_async(std::function<void()> task) {
 
 static Status collect_garbage_files(const TabletMetadataPB& metadata, const std::string& base_dir,
                                     AsyncFileDeleter* deleter, AsyncBundleFileDeleter* bundle_file_deleter,
-                                    int64_t* garbage_data_size) {
+                                    int64_t* garbage_data_size, const TabletRetainInfo& retain_info) {
     for (const auto& rowset : metadata.compaction_inputs()) {
+        if (retain_info.contains_rowset(rowset.id())) {
+            continue;
+        }
         for (const auto& segment : rowset.segments()) {
             if (rowset.bundle_file_offsets_size() > 0 && bundle_file_deleter != nullptr) {
                 RETURN_IF_ERROR(bundle_file_deleter->delete_file(join_path(base_dir, segment)));
@@ -334,6 +338,9 @@ static Status collect_garbage_files(const TabletMetadataPB& metadata, const std:
         *garbage_data_size += rowset.data_size();
     }
     for (const auto& file : metadata.orphan_files()) {
+        if (retain_info.contains_file(file.name())) {
+            continue;
+        }
         RETURN_IF_ERROR(deleter->delete_file(join_path(base_dir, file.name())));
         *garbage_data_size += file.size();
     }
@@ -384,7 +391,8 @@ static Status collect_files_to_vacuum(TabletManager* tablet_mgr, std::string_vie
                                       VacuumTabletMetaVerionRange* vacuum_version_range,
                                       AsyncFileDeleter* datafile_deleter, AsyncFileDeleter* metafile_deleter,
                                       AsyncBundleFileDeleter* bundle_file_deleter, int64_t* total_datafile_size,
-                                      int64_t* vacuumed_version, int64_t* extra_datafile_size) {
+                                      int64_t* vacuumed_version, int64_t* extra_datafile_size,
+                                      const TabletRetainInfo& retain_info) {
     auto t0 = butil::gettimeofday_ms();
     auto meta_dir = join_path(root_dir, kMetadataDirectoryName);
     auto data_dir = join_path(root_dir, kSegmentDirectoryName);
@@ -412,7 +420,7 @@ static Status collect_files_to_vacuum(TabletManager* tablet_mgr, std::string_vie
             if (skip_check_grace_timestamp) {
                 DCHECK_LE(version, final_retain_version);
                 RETURN_IF_ERROR(collect_garbage_files(*metadata, data_dir, datafile_deleter, bundle_file_deleter,
-                                                      &prepare_vacuum_file_size));
+                                                      &prepare_vacuum_file_size, retain_info));
             } else {
                 int64_t compare_time = 0;
                 if (metadata->has_commit_time() && metadata->commit_time() > 0) {
@@ -448,7 +456,7 @@ static Status collect_files_to_vacuum(TabletManager* tablet_mgr, std::string_vie
 
                     // The metadata will be retained, but garbage files recorded in it can be deleted.
                     RETURN_IF_ERROR(collect_garbage_files(*metadata, data_dir, datafile_deleter, bundle_file_deleter,
-                                                          total_datafile_size));
+                                                          total_datafile_size, retain_info));
                 } else {
                     DCHECK_LE(version, final_retain_version);
                     final_retain_version = version;
@@ -473,6 +481,9 @@ static Status collect_files_to_vacuum(TabletManager* tablet_mgr, std::string_vie
     DCHECK_LE(version, final_retain_version);
     if (vacuum_version_range == nullptr) {
         for (auto v = version + 1; v < final_retain_version; v++) {
+            if (retain_info.contains_version(v)) {
+                continue;
+            }
             RETURN_IF_ERROR(metafile_deleter->delete_file(join_path(meta_dir, tablet_metadata_filename(tablet_id, v))));
         }
     } else {
@@ -499,7 +510,8 @@ static void erase_tablet_metadata_from_metacache(TabletManager* tablet_mgr, cons
 static Status vacuum_tablet_metadata(TabletManager* tablet_mgr, std::string_view root_dir,
                                      std::vector<TabletInfoPB>& tablet_infos, int64_t min_retain_version,
                                      int64_t grace_timestamp, bool enable_file_bundling, int64_t* vacuumed_files,
-                                     int64_t* vacuumed_file_size, int64_t* vacuumed_version, int64_t* extra_file_size) {
+                                     int64_t* vacuumed_file_size, int64_t* vacuumed_version, int64_t* extra_file_size,
+                                     const std::unordered_set<int64_t>& retain_versions) {
     DCHECK(tablet_mgr != nullptr);
     DCHECK(std::is_sorted(tablet_infos.begin(), tablet_infos.end(),
                           [](const auto& a, const auto& b) { return a.tablet_id() < b.tablet_id(); }));
@@ -519,13 +531,16 @@ static Status vacuum_tablet_metadata(TabletManager* tablet_mgr, std::string_view
     int64_t final_vacuum_version = std::numeric_limits<int64_t>::max();
     int64_t max_vacuum_version = 0;
     for (auto& tablet_info : tablet_infos) {
+        TabletRetainInfo tablet_retain_info;
+        RETURN_IF_ERROR(tablet_retain_info.init(tablet_info.tablet_id(), retain_versions, tablet_mgr));
+
         int64_t tablet_vacuumed_version = 0;
         AsyncFileDeleter datafile_deleter(config::lake_vacuum_min_batch_delete_size);
         AsyncFileDeleter metafile_deleter(INT64_MAX, metafile_delete_cb);
         RETURN_IF_ERROR(collect_files_to_vacuum(tablet_mgr, root_dir, tablet_info, grace_timestamp, min_retain_version,
                                                 vacuum_version_range.get(), &datafile_deleter, &metafile_deleter,
                                                 &bundle_file_deleter, vacuumed_file_size, &tablet_vacuumed_version,
-                                                extra_file_size));
+                                                extra_file_size, tablet_retain_info));
         RETURN_IF_ERROR(datafile_deleter.finish());
         (*vacuumed_files) += datafile_deleter.delete_count();
         if (!enable_file_bundling) {
@@ -558,6 +573,9 @@ static Status vacuum_tablet_metadata(TabletManager* tablet_mgr, std::string_view
             }
         }
         for (auto v = vacuum_version_range->min_version; v < vacuum_version_range->max_version; v++) {
+            if (retain_versions.find(v) != retain_versions.end()) {
+                continue;
+            }
             RETURN_IF_ERROR(metafile_deleter.delete_file(join_path(meta_dir, tablet_metadata_filename(0, v))));
         }
         RETURN_IF_ERROR(metafile_deleter.finish());
@@ -645,6 +663,10 @@ Status vacuum_impl(TabletManager* tablet_mgr, const VacuumRequest& request, Vacu
     auto min_retain_version = request.min_retain_version();
     auto grace_timestamp = request.grace_timestamp();
     auto min_active_txn_id = request.min_active_txn_id();
+    std::unordered_set<int64_t> retain_versions;
+    if (request.retain_versions_size() > 0) {
+        retain_versions.insert(request.retain_versions().begin(), request.retain_versions().end());
+    }
 
     int64_t vacuumed_files = 0;
     int64_t vacuumed_file_size = 0;
@@ -656,7 +678,7 @@ Status vacuum_impl(TabletManager* tablet_mgr, const VacuumRequest& request, Vacu
 
     RETURN_IF_ERROR(vacuum_tablet_metadata(tablet_mgr, root_loc, tablet_infos, min_retain_version, grace_timestamp,
                                            request.enable_file_bundling(), &vacuumed_files, &vacuumed_file_size,
-                                           &vacuumed_version, &extra_file_size));
+                                           &vacuumed_version, &extra_file_size, retain_versions));
     extra_file_size -= vacuumed_file_size;
     if (request.delete_txn_log()) {
         RETURN_IF_ERROR(vacuum_txn_log(root_loc, min_active_txn_id, &vacuumed_files, &vacuumed_file_size));
@@ -957,7 +979,7 @@ Status delete_tablets_impl(TabletManager* tablet_mgr, const std::string& root_di
                                               can_bundle_meta_file_to_be_deleted(versions_and_states[garbage_version])
                                                       ? nullptr
                                                       : &dummy_bundle_file_deleter,
-                                              &dummy_file_size));
+                                              &dummy_file_size, TabletRetainInfo()));
                 if (metadata->has_prev_garbage_version()) {
                     garbage_version = metadata->prev_garbage_version();
                 } else {
